requirements.txt
Django==3.2
djangorestframework>=3.12.4,<3.13
psycopg2-binary==2.9.5
drf-spectacular>=0.15.1,<0.16
